ext4: re-inline ext4_rec_len_(to|from)_disk functions
Patchwork-id: 27386
O-Subject: [PATCH] ext4: re-inline ext4_rec_len_(to|from)_disk functions
Bugzilla: 522808
RH-Acked-by: Ric Wheeler <rwheeler@redhat.com>
RH-Acked-by: Christoph Hellwig <chellwig@redhat.com>
RH-Acked-by: Jeff Moyer <jmoyer@redhat.com>
RH-Acked-by: Edward Shishkin <edward@redhat.com>

This is for bug #522808
[6.0] I/O performance for ext4 file creation is inferior to ext3's

I've not sent this upstream yet, but will shortly. I already have
nominal agreement that this is worth fixing.  Patch below
is the least-impact solution for this stage of the release.

commit 3d0518f4, ext4: New rec_len encoding for very
large blocksizes made several changes to this path, but from
a perf perspective, un-inlining ext4_rec_len_from_disk() seems
most significant.  This function is called from ext4_check_dir_entry(),
which on a file-creation workload is caled extremely often.

I tested this with bonnie:

# bonnie++ -u root -s 0 -f -x 200 -d /mnt/test -n 32

(this does 200 iterations) and got this for the file creations:

ext4 stock:   Average =  21206.8 files/s
ext4 inlined: Average =  22346.7 files/s  (+5%)

We may get a little more by optimizing around the extra complexity
for > 64K blocks, but for now simply inlining is a nice improvement.

Signed-off-by: Eric Sandeen <sandeen@redhat.com>