mptfusion: Corrected declaration of device_missing_delay
Patchwork-id: 27390
O-Subject: [RHEL6 PATCH3/7] mptfusion: Corrected declaration of
	device_missing_delay
Bugzilla: 618625
RH-Acked-by: Rob Evers <revers@redhat.com>

device missing delay is 8 bit value in io unit pg1. Making correct variable
declaration for device_missing_delay.

The driver is storing the calculated device missing delay in IOC structure
as a u8 instead of a u16. It needs to be a u16 if the delay is > 255.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>