xfs: don't walk AGs that can't hold inodes
Patchwork-id: 27399
O-Subject: [RHEL6 PATCH] xfs: don't walk AGs that can't hold inodes
Bugzilla: 621044
RH-Acked-by: Eric Sandeen <sandeen@redhat.com>
RH-Acked-by: Ric Wheeler <rwheeler@redhat.com>

BZ: 621044

This fixes a regression introduced in the backport of the fix for #617035.
The upstream code no longer does conditional initialisation of the per-ag inode
trees, whilst the RHEL6 code does. Hence the new AG-walking code backported in
617035 does not check the pag->pag_ici_init flag and can walk into AGs with
uninitialised inode cache locks and trees.

There is no upstream commit as the bug is specific to the RHEL6 tree.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>