mptfusion: fix DMA boudary
Patchwork-id: 27440
O-Subject: Re: [RHEL6 PATCH0/7] mptfusion: update to 3.04.16 critical fixes
Bugzilla: 618625
RH-Acked-by: Jeff Moyer <jmoyer@redhat.com>

This is for bz#618625

lsi:"Earlier somewhere 2.6.26 time kernel always had 512 byte Dma alignment by
default. It means all those kernel was considering 512 byte boundary for both
SAS and SATA drives. when there was a change in scsi mid layer (4 byte boundary
instead of 512 byte) we have to move 512 boundary alignment to mptfusion
driver."

This patch is what Kashyap (LSI) acked on upstream.
The other version:
+	if (rphy && (rphy->identify.target_port_protocols & SAS_PROTOCOL_SATA))
+		blk_queue_dma_alignment(sdev->request_queue, 512 - 1);
which changes the boundary only for SATA transfers is not wrong,
but it is better to use what is going into upstream.

Tomas

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>