Skip depmod when installing to non-standard INSTALL_MOD_PATH
Patchwork-id: 27498
O-Subject: [RHEL6 PATCH] [modules] BZ 609170: Skip depmod when installing to
	non-standard INSTALL_MOD_PATH
Bugzilla: 609170
RH-Acked-by: Amerigo Wang <amwang@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

https://bugzilla.redhat.com/show_bug.cgi?id=609170

The Linux kernel build system (Kbuild) includes a convenience call to
depmod. Distributions such as Red Hat Enterprise Linux 6 already perform
their own calls to depmod during kernel installation, and the additional
call is thus a harmless benefit for those building additional modules. A
problem arises when an assumption is made that depmod will always be
run, especially when installing to non-standard module directories. In
this case, running depmod is non-sensical anyway. In the worst case,
it can result in additional depmod files in that non-standard path.

This patch removes calls to depmod in the case that modules are being
installed into a non-standard INSTALL_MOD_PATH, such as during the
build of external kernel module packages. When performing such builds,
we do not want to create extraneous depmod files that may end up
replacing the standard system-installed and generated files.

Signed-off-by: Jon Masters <jcm@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>