radeon: Don't limit vram size to aperture size
Patchwork-id: 27513
O-Subject: [PATCH] [RHEL 6.0 PATCH] radeon: Don't limit vram size to aperture
	size
Bugzilla: 622039
RH-Acked-by: Adam Jackson <ajax@redhat.com>

rhbz#: 622039

Right now we limit the radeon memory controller's accessible space to the
size of the aperture. If there's video ram above that address then the
firmware may not be able to access the GPU scratch registers, and if the
firmware uses those registers in order to perform brightness transitions
then the hardware may lock up in response to backlight changes. Lift this
restriction in order to avoid hanging machines.

Reviewed-by: Jerome Glisse <jglisse@redhat.com>

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>