be2net: maintain multicast packet count in driver
Patchwork-id: 27521
O-Subject: [RHEL6.0 PATCH] be2net: maintain multicast packet count in driver
Bugzilla: 621287
RH-Acked-by: Andy Gospodarek <gospo@redhat.com>

BZs:
#621287 - be2net: wrong packet statistics cause net-snmp error messages every 15s

Description:
This patch fixes the situation when the multicast packets counter is
greater than total Rx packet counter. The problem is total Rx counter is
maintained by the driver and is initilized to zero when the driver is
loaded. For the multicast packets is used the H/W counter and this is
zeroed from the point the H/W is reset. The solution is to count multicast
packets also in the driver.

Upstream status:
Not yet, will be soon.

Test status:
Tested by myself.

Signed-off-by: Ivan Vecera <ivecera@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>