increase flush timeout
Patchwork-id: 27524
O-Subject: [RHEL 6 PATCH] scsi: increase flush timeout
Bugzilla: 605322
RH-Acked-by: Rob Evers <revers@redhat.com>
RH-Acked-by: Robert S Peterson <rpeterso@redhat.com>
RH-Acked-by: Tomas Henzl <thenzl@redhat.com>
RH-Acked-by: Ric Wheeler <rwheeler@redhat.com>

From: Mike Christie <mchristi@redhat.com>

This is for bz 605322.

We have been seeing the flush request timeout on several
different setups. After several tries upstrem where I
made it configurable (http://marc.info/?l=linux-scsi&m=127957359200466&w=2
- block layer maintainer did not like this because he through
it would be difficult for users to config and they would
not know to set it until they have got IO errors which
would be too late)
and added a long timeout (http://marc.info/?l=linux-scsi&m=128086333510289&w=2
 - this timeout is just too long and the scsi maintainer
and Ric said that if it is taking that long then it is
going to cause the hung timer to fire constantly and
the user needs to try something else like changing the cache settings),
so in the end we agreed to just increase the hard coded value to 60 secs.

I have sent the patch below here
http://marc.info/?l=linux-scsi&m=128154253524611&w=2
It is not yet merged in upstream, but I feel it is safe regression
wise because it is just increasing a hard coded value, and it is safe
upstream wise (scsi and block layer maintainer agree on approach now,
but this exact patch has not been merged).

Without the change, users could experience writes that do not
make it on disk (cache sync is used to tell device to write
from cache to disk) and a slow down in IO because the scsi cmd
timer will run and start the scsi eh.

I have tested the patch here, and it fixes the problem in the bz.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>