Revert: qla2xxx: Propogate transport disrupted status for cable pull conditions for faster failover
Patchwork-id: 27525
O-Subject: [RHEL 6 PATCH] Revert: qla2xxx: Propogate transport disrupted status
	for cable pull conditions for faster failover
Bugzilla: 622041
RH-Acked-by: Mike Christie <mchristi@redhat.com>
RH-Acked-by: Rob Evers <revers@redhat.com>

Bugzilla
========

622041

Description
===========

This patch is a fix to a regression in which a short switch port offline/online
will cause the SCSI midlayer to exhaust all of its retries.  The reason for this
is that the current return value of DID_TRANSPORT_DISRUPTED will cause the SCSI
mid-layer to immediately retry as opposed to the old semantic of returning
DID_RESET.

This patch reverts back to the old semanitc of just returning DID_RESET.  It is
a revert of the commit:

qla2xxx: Propogate transport disrupted status for cable pull conditions for faster failover

Upstream Status
===============

The current plan is to fix the semantics around DID_TRANSPORT_DISRUPTED and
repost that fix for inclusion in the upstream kernel.  Once is has been posted
upstream it would also be recommended that it be included in a RHEL 6 z-stream
kernel.

Testing
=======

This patch was created and sanity checking performed by Mike Christie.  IBM has
already verified that returning DID_RESET fixes this issue.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>