libfc: call fc_remote_port_chkready under the host lock
Patchwork-id: 27527
O-Subject: [RHEL 6 PATCH]: libfc: call fc_remote_port_chkready under the host
	lock.
Bugzilla: 623786
RH-Acked-by: Tomas Henzl <thenzl@redhat.com>
RH-Acked-by: Stanislaw Gruszka <sgruszka@redhat.com>
RH-Acked-by: Rob Evers <revers@redhat.com>

From: Mike Christie <mchristi@redhat.com>

This is for bz 623786.

The rport port state and flags are set under the host lock,
so this patch calls fc_remote_port_chkready with the host lock
held like is also done in the other fc drivers.

This fixes a bug where we can race with the thread
that removes the rport and IO will get run
(and failed) instead of requeued. If this happens enough
times it will get failed to the app.

Intel has been testing our code and needs this for EMC qualification.
I did some tests and was able to hit the problem (I think - it
is such a small race it is very hard for me to hit) and
this has fixed it for me.

This only affects the fcoe.ko driver from intel. It will not
affect fnic (the other libfc user), because they implement
their own queuecommand function.

Patch is in scsi-misc tree:
http://git.kernel.org/?p=linux/kernel/git/jejb/scsi-misc-2.6.git;a=commit;h=be948fc30dabc347e4fba4daf4da9fd83d2abf49

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>