O_DIRECT: fix the splitting up of contiguous I/O
Patchwork-id: 27530
O-Subject: [RHEL6 patch] O_DIRECT: fix the splitting up of contiguous I/O
Bugzilla: 622504
RH-Acked-by: Josef Bacik <josef@redhat.com>
RH-Acked-by: Jerome Marchand <jmarchan@redhat.com>

Hi,

Upstream commit c2c6ca4 (direct-io: do not merge logically
non-contiguous requests) introduced a bug whereby all O_DIRECT I/Os were
submitted a page at a time to the block layer.  The problem is that the
code expected dio->block_in_file to correspond to the current page in
the dio.  In fact, it corresponds to the previous page submitted via
submit_page_section.  This was purely an oversight, as the
dio->cur_page_fs_offset field was introduced for just this purpose.
This patch simply uses the correct variable when calculating whether
there is a mismatch between contiguous logical blocks and contiguous
physical blocks (as described in the comments).

I also switched the if conditional following this check to an else if,
to ensure that we never call dio_bio_submit twice for the same dio (in
theory, this should not happen, anyway).

I've tested this by running blktrace and verifying that a 64KB I/O was
submitted as a single I/O.  I also ran the patched kernel through
xfstests' aio tests using xfs, ext4 (with 1k and 4k block sizes) and
btrfs and verified that there were no regressions as compared to an
unpatched kernel.

All of the testing was done under an upstream kernel.  This addresses
bug 622504.  Comments, as always, are welcome.

Cheers,
Jeff

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>