ext4: fix discard failure detection
Patchwork-id: 27557
O-Subject: [PATCH RHEL6] ext4: fix discard failure detection
Bugzilla: 608731
RH-Acked-by: Jeff Moyer <jmoyer@redhat.com>
RH-Acked-by: Mike Snitzer <snitzer@redhat.com>

This is for bug 608731 - ext4: stop issuing discards if not supported by device

(which is ON_QA ... do we need a new bug to fix the fix?)

QA caught this in testing; total thinko on my part.  sb_issue_discard
returns a negative error but we were checking positive.

Flipping the sign is all that is needed to detect the failed
discard.

This one isn't a terrible problem (issuing discards only happen
if you mount with that option, and continuing to issue discards
to a device which rejects them is just a little overhead) but it's
a simple fix.

Thanks,
-Eric

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>