nfsd4: fix delegation recall race use-after-free
Patchwork-id: 27571
O-Subject: [PATCH 10/23] nfsd4: fix delegation recall race use-after-free
Bugzilla: 624081 637278
RH-Acked-by: Rik van Riel <riel@redhat.com>

From: J. Bruce Fields <bfields@citi.umich.edu>

When the rarely-used callback-connection-changing setclientid occurs
simultaneously with a delegation recall, we rerun the recall by
requeueing it on a workqueue.  But we also need to take a reference on
the delegation in that case, since the delegation held by the rpc itself
will be released by the rpc_release callback.

Signed-off-by: J. Bruce Fields <bfields@citi.umich.edu>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>