nfsd4: fix openmode checking on IO using lock stateid
Patchwork-id: 27574
O-Subject: [PATCH 18/23] nfsd4: fix openmode checking on IO using lock stateid
Bugzilla: 624081
RH-Acked-by: Rik van Riel <riel@redhat.com>

From: J. Bruce Fields <bfields@redhat.com>

It is legal to perform a write using the lock stateid that was
originally associated with a read lock, or with a file that was
originally opened for read, but has since been upgraded.

So, when checking the openmode, check the mode associated with the
open stateid from which the lock was derived.

Signed-off-by: J. Bruce Fields <bfields@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>