SUNRPC: prevent task_cleanup running on freed xprt
Patchwork-id: 27614
O-Subject: [PATCH 38/43] SUNRPC: prevent task_cleanup running on freed xprt
Bugzilla: 624069
RH-Acked-by: Rik van Riel <riel@redhat.com>

From: J. Bruce Fields <bfields@redhat.com>

We saw a report of a NULL dereference in xprt_autoclose:

	https://bugzilla.redhat.com/show_bug.cgi?id=611938

This appears to be the result of an xprt's task_cleanup running after
the xprt is destroyed.  Nothing in the current code appears to prevent
that.

Signed-off-by: J. Bruce Fields <bfields@redhat.com>
Signed-off-by: Trond Myklebust <Trond.Myklebust@netapp.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>