ext4: fix NULL pointer dereference in tracing
Patchwork-id: 27631
O-Subject: [RHEL6 PATCH] ext4: fix one more tracing oops
Bugzilla: 619013
RH-Acked-by: Dave Chinner <dchinner@redhat.com>
RH-Acked-by: Ric Wheeler <rwheeler@redhat.com>
RH-Acked-by: Josef Bacik <josef@redhat.com>
RH-Acked-by: Jeff Moyer <jmoyer@redhat.com>

This is an update patch for bug 619013 -
ext4 trace events cause NULL pointer dereferences

An earlier patch got sent for one problem, but there was another
trace path which could oops, and it was not fixed.

ext4_mb_release_group_pa() does:

        if (ac) {
                ac->ac_sb = sb;
                ac->ac_inode = NULL;
                ac->ac_b_ex.fe_group = group;
                ac->ac_b_ex.fe_start = bit;
                ac->ac_b_ex.fe_len = pa->pa_len;
                ac->ac_b_ex.fe_logical = 0;
                trace_ext4_mballoc_discard(ac);
        }

and the tracepoint does:

        TP_fast_assign(
                __entry->dev            = ac->ac_inode->i_sb->s_dev;
                __entry->ino            = ac->ac_inode->i_ino;

thereby dereferencing the ac_inode we just set to NULL.

ext4 tracing doesn't do this upstream, so this is a RHEL6-only patch.
I tested this by enabling all tracepoints in ext4 and running through
the generic xfstests; without the patch it oopses reliably, with the
patch we survive.

Thanks,
-Eric

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>