ext4: consolidate in_range() definitions
Patchwork-id: 27632
O-Subject: [RHEL6 PATCH] ext4: consolidate in_range() definitions
Bugzilla: 621829
RH-Acked-by: Ric Wheeler <rwheeler@redhat.com>
RH-Acked-by: Josef Bacik <josef@redhat.com>

This is for bug 621829 -
[6.0] a panic happens on ext4('kernel BUG at fs/ext4/mballoc.c:2958!')

There is an overflow in the extent range checking used in
fs/ext4/extents.c; this overflow can lead to an oops with the
reproducer attached to the bug.

The upstream patch below resolves it by replacing the open-coded
range checking with in_range(), which avoids the overflow by
changing essentially "a < b" to "a <= b - 1"

In other words; we had:

if (block >= cex->ec_block && block < cex->ec_block + cex->ec_len) {

ec_block is u32; ec_len is u16; maximal block and len 1
would lead to 4294967295 + 1 = 0, and the test failed when it
should not have.

The macro definition does:

#define in_range(b, first, len)	((b) >= (first) && (b) <= (first) + (len) - 1)

so we get this sort of test:

if (block >= cex->ec_block && block <= cex->ec_block + cex->ec_len - 1) {

which takes care of the overflow, and fixes the testcase. When
we're not at the u32 limits, the test is equivalent.  I've also
done the standard 70-test xfstests run with this change in place.

Thanks,
-Eric

commit 731eb1a03a8445cde2cb23ecfb3580c6fa7bb690
Author: Akinobu Mita <akinobu.mita@gmail.com>
Date:   Wed Mar 3 23:55:01 2010 -0500

    ext4: consolidate in_range() definitions

    There are duplicate macro definitions of in_range() in mballoc.h and
    balloc.c.  This consolidates these two definitions into ext4.h, and
    changes extents.c to use in_range() as well.

    Signed-off-by: Akinobu Mita <akinobu.mita@gmail.com>
    Signed-off-by: "Theodore Ts'o" <tytso@mit.edu>
    Cc: Andreas Dilger <adilger@sun.com>