Fix the kdump kernel OOMs caused by passthrough mode setting
Patchwork-id: 27664
O-Subject: [RHEL6 PATCH] Fix the kdump kernel OOMs caused by passthrough mode
	setting
Bugzilla: 624059
RH-Acked-by: Christopher Lalancette <clalance@redhat.com>
RH-Acked-by: Vivek Goyal <vgoyal@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

Resolves BZ 624059

This patch was developed to fix the RHEL6 issue introduced by AMD passthrough mode
patch.   It checks the max_pfn and only sets swiotlb if the max_pfn > MAX_DMA32_PFN.
In the kdump kernel max_pfn seems to point to the end of the kdump memory which is
below 4gb and thus swiotlb is now 0 and the 64M are not longer allocated at boot
of the kdump kernel.

Fixes the OOM issues seen with kdump kernels.

Testing done successfully on Dinar system, the OOM kill was fixed with this patch.

Please review and ACK.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>