Force "pci=noacpi" on HP xw9300
Patchwork-id: 27666
O-Subject: [RHEL6 BZ 615276 PATCH]: Force "pci=noacpi" on HP xw9300
Bugzilla: 615276
RH-Acked-by: Matthew Garrett <mjg@redhat.com>
RH-Acked-by: Tomas Henzl <thenzl@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

The kernel fails to boot when enabling "ACPI PCI segments" in the BIOS of the
xw9300.  This happens because the ACPI functions enabled by the BIOS option
are incorrect and the result is a series of very loud ACPI warnings.

Force "pci=noacpi" on the xw9300.

The xw9300 will then boot with one PCI domain.

Successfully tested by me.

Resolves BZ 615276.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>