Run EFI in physical mode to enable kdump on EFI-booted system
Patchwork-id: 27672
O-Subject: [RHEL6.0 PATCH v2] Run EFI in physical mode to enable kdump on
	EFI-booted system
Bugzilla: 593111
RH-Acked-by: Vivek Goyal <vgoyal@redhat.com>
RH-Acked-by: Rik van Riel <riel@redhat.com>

Hi all,

Updated patch for bz593111
https://bugzilla.redhat.com/show_bug.cgi?id=593111

I posted a patch to upstream and got some comments, so I updated it. The
upstream thread is below.
http://marc.info/?t=128172737700001&r=1&w=2
http://marc.info/?t=128200019900002&r=1&w=2

Basically this is almost the same as old version, just changed some variables
to per-CPU.

brewbuild is ongoing.
http://brewweb.devel.redhat.com/brew/taskinfo?taskID=2686063

[Description]
Current kdump cannot work on EFI-booted system because of two problems. One
problem is that 2nd kernel cannot find ACPI Table(RSDP). Other problem is that
EFI system table is changed by 1st kernel. The URL below is explaining the
detail.
http://marc.info/?l=linux-kernel&m=128018221820234&w=2
In this thread, the idea to run EFI in physical mode was proposed. The
attached patch implements it.

Basically EFI is in physical mode at first and it's switched to virtual
mode after calling SetVirtualAddressMap. By applying this patch, you can
run EFI always in physical mode. And you can also specify "virtefi" as
kernel boot parameter to run EFI in virtual mode as before. Note that
this patch supports only x86_64.

Basic idea of this patch is to create EFI own pagetable. This pagetable maps
physical address of EFI runtime to the virtual address which is the same value
so that we can call it directly. For example, physical address 0x800000 is
mapped to virtual address 0x800000. Before calling EFI runtime, cr3 register
is switched to this pagetable, and restored when we come back from EFI.

Thanks,
Takao Indoh

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>