nv50: insert a delay before fb change to prevent display engine hang
Patchwork-id: 27698
O-Subject: [RHEL6 PATCH] drm/nv50: insert a delay before fb change to prevent
	display engine hang
Bugzilla: 618225

rhbz#618225

The display engine on this card appears to strongly dislike when we submit
multiple framebuffer-only display switches too soon after each other (as in the
dual-head case here, when switching to/from X and a VT), and gets stuck in a
state where it will no longer accept additional updates.

The first solution, and what will likely be done upstream eventually, is to
insert another UPDATE command into the display push buffer right before we
do the framebuffer state changes.  This works just fine and seems to keep
the display engine happy.

Being paranoid for EL6 at this late a stage, I've replaced this with a delay on
the CPU to prevent any unforeseen side-effects of modifying the sequences of
commands we issue to the display engine.

As indicated above, this patch is not upstream.  However, the original
solution mentioned above will likely make it upstream once additional
testing has been done.

Signed-off-by: Ben Skeggs <bskeggs@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>