acpi: Update battery information on notification 0x81
Patchwork-id: 27721
O-Subject: [PATCH] [RHEL 6.0 PATCH] acpi: Update battery information on
	notification 0x81
Bugzilla: 606388
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

rhbz #606388

A notification event 0x81 from an ACPI battery device requires us to
re-read the battery information structure. Do so, and if the battery's
reporting units have changed (as is the case on some Thinkpads) destroy
and recreate the battery in order to populate the fields correctly.

Upstream status: Submitted, not yet committed

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>