ext4: fix one more tracing oops
Patchwork-id: 27724
O-Subject: [PATCH RHEL6] ext4: fix one more tracing oops
Bugzilla: 619013
RH-Acked-by: Josef Bacik <josef@redhat.com>
RH-Acked-by: Robert S Peterson <rpeterso@redhat.com>
RH-Acked-by: Valerie Aurora <vaurora@redhat.com>

This is for Bug 619013 again :(

The first patch sent for this bug had a bug of its own; we test whether
ac->inode is NULL, but it's not always initialized.  So if we get non-null
garbage, dereferencing ac_inode->i_ino still goes boom.  (thanks to Josef
for spotting that!)

This happens down the ext4_mb_release_group_pa path, when we never have
useful information in the ac_inode -anyway- so just drop that bit of tracing.

(In this case the allocation context was allocated only as a container
to hold relevant tracing bits, and the inode is never filled in)

Tested (again) by running through xfstests with tracing enabled,
this time with CONFIG_SLAB_DEBUG turned on so that freshly-allocated
but uninitialized allocation contexts have poison in the inode ptr.
That reproduces the latest reported bug, and with this patch it succeeds.

I'll send this patch upstream today.

Signed-off-by: Eric Sandeen <sandeen@redhat.com>