acpi: Update battery information on notification 0x81
Patchwork-id: 27725
O-Subject: [PATCH] [RHEL 6.0 PATCH v2] acpi: Update battery information on
	notification 0x81
Bugzilla: 606388
RH-Acked-by: Dean Nelson <dnelson@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

rhbz #606388

A notification event 0x81 from an ACPI battery device requires us to
re-read the battery information structure. Do so, and if the battery's
reporting units have changed (as is the case on some Thinkpads) destroy
and recreate the battery in order to populate the fields correctly.

Fixed a potential "Declared but not used" issue and fix a test for quality
rather than assignment.

Upstream status: Submitted, not yet committed

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>