kabi: Remove stale symbols from kABI lists
Patchwork-id: 27753
O-Subject: [RHEL6 PATCH 1/2] kabi: BZ 625217: Remove stale symbols from kABI
	lists
Bugzilla: 625217

Since we began working on the most recent kABI updates, some kABI symbols
have been removed from the kernel due to code changes. Of these, the
three blk_queue_* symbols have been directly replaced (see next patch),
whereas the others are gone and have no direct replacement. This is true
for example of kmalloc_caches, and of the defunct IRDA symbols.

Here is the full list of symbols that are stale and will be removed
after applying this patch, for the sake of completeness (note that
some are arch-specific, on PPC64 and s390x, whereas some do not
exist on all architectures, and the warn_slowpath_* symbols
are only removed on s390x systems and not on others):

        - alloc_irdadev
        - async_unwrap_char
        - async_wrap_skb
        - blk_queue_max_hw_segments
        - blk_queue_max_phys_segments
        - blk_queue_max_sectors
        - irda_device_set_media_busy
        - irda_init_max_qos_capabilies
        - irda_qos_bits_to_value
        - irda_register_dongle
        - irda_unregister_dongle
        - irlap_close
        - irlap_open
        - kmalloc_caches
        - ppc_pci_io
        - sirdev_set_dtr_rts
        - skb_dma_map
        - skb_dma_unmap
        - warn_slowpath_fmt
        - warn_slowpath_null

This patch will allow us to clean up the final kABI whitelists.

Signed-off-by: Jon Masters <jcm@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>