cifs: fix NULL pointer dereference in cifs_find_smb_ses
Patchwork-id: 27789
O-Subject: [RHEL6 PATCH] BZ#625583: cifs: fix NULL pointer dereference in
	cifs_find_smb_ses
Bugzilla: 625583
RH-Acked-by: Jeff Moyer <jmoyer@redhat.com>
RH-Acked-by: Sachin Prabhu <sprabhu@redhat.com>

cifs_find_smb_ses assumes that the vol->password field is a valid
pointer, but that's only the case if a password was passed in via
the options string. It's possible that one won't be if there is
no mount helper on the box.

Reported-by: diabel <gacek-2004@wp.pl>
Signed-off-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: Steve French <sfrench@us.ibm.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>