add quirk to fixup mcp55 interrupt routing to let kdump work
Patchwork-id: 27834
O-Subject: Re: [RHEL6.1 PATCH] x86: add quirk to fixup mcp55 interrupt routing
	to let kdump work (bz 562134)
Bugzilla: 562134
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Vivek Goyal <vgoyal@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Hey all-
	Back in RHEL5, I fixed a problem for LLNL (bz 462915) in which we had to
flip a secret register on the nvidia mcp55 northbridge chip so that interrupts
could be broadcast to all processors on a hypertransport bus, rather than only
the bsp.  The plan was to rewrite the apic code so that it was initalized
earlier so that we didn't boot the kernel in legacy mode.  Unfortunately, the
invasiveness of that change (needed to initalize acpi earlier as well, and make
some other major changes), led to that work never getting done upstream.  And it
was recently noted that the same kdump problem exists on opteron systems with
mcp55 chips.  Since mcp55 is the only northbridge that seems to participate in
this sillyness, I'm just forward porting the RHEL5 quirk I added to allow kdump
to work on those systems.  Tested by pingtan and confirmed to fix bz 562134

Neil

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>