s390x: kdump: allow zfcpdump to mount and write to ext4 file systems
Patchwork-id: 27982
O-Subject: [RHEL6 PATCH 5/5] [s390x] kdump: allow zfcpdump to mount and write to
	ext4 file systems
Bugzilla: 628676
RH-Acked-by: Neil Horman <nhorman@redhat.com>

Description
-----------
The zfcpdump is not able to mount ext4 file systems.  Because ext4
is the default file system on RHEL6, also add respective support
to the zfcpdump application.  Switch on ext4 in the s390x kernel config.

Bugzilla
--------
BZ 628676
https://bugzilla.redhat.com/show_bug.cgi?id=628676

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>