UV: use virtual efi on SGI systems
Patchwork-id: 28192
O-Subject: [Patch RHEL6.0.Z] UV: use virtual efi on SGI systems
Bugzilla: 627653
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Neil Horman <nhorman@redhat.com>
RH-Acked-by: Vivek Goyal <vgoyal@redhat.com>

BZ#627653 RHEL-UV: kernel panic on boot uvsw-sys
First post.

The following patch sets usevirtefi on SGI systems.  This is called early
enough that EFI runs in virtual mode.  Because virtual is the upstream
default and this was changed very late in Rhel6.0 we have zero exposure
running EFI in physical mode, which includes bios runtime code.

We know the bios routines work when run in virtual mode, but they are
known to fail in physical mode with the existing rhel6 kernel and with
a patch proposed by Neil Horman (see the BZ).

This patch is known to work on machines inside SGI including the
large machine we make available to engineering.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>