fix split_huge_page error like mapcount 3 page_mapcount 2
Patchwork-id: 28213
O-Subject: [PATCH RHEL6] fix split_huge_page error like mapcount 3
	page_mapcount 2
Bugzilla: 622327 640611
RH-Acked-by: Rik van Riel <riel@redhat.com>
RH-Acked-by: Johannes Weiner <jweiner@redhat.com>
RH-Acked-by: Larry Woodman <lwoodman@redhat.com>
RH-Acked-by: Marcelo Tosatti <mtosatti@redhat.com>

This shall apply to all RHEL6 kernels (6.0 updates and 6.1).

https://bugzilla.redhat.com/show_bug.cgi?id=622327

While debugging the 622327 I created some evil application flooding
the kernel with forks and split_huge_page in parallel that can
reproduce this BUG_ON instantly (not divulging it but ask if you need
it for something). With this patch system is up for 11 hours without
problems.

It'd be possible to fix it in other ways (i.e. changing
split_huge_page not to be strict anymore and ignore huge-pmd that have
already the splitting bit set) but the root cause of the bug is the
below, and this makes the anon-vma structures pointing to the right
"mm" place, also during fork so fixing the bug completely.

It's the very same BUG_ON as in 622327 that I managed to reproduce
with my evil app, but I'm not too optimistic that this will also close
622327 (622327 is like mapcount 2 page_mapcount 3, here I got the
mapcount 3 and page_mapcount 2...). But I'll build a kernel to test
just in case the bug has side effects that I can't easily
foresee. Still working on 622327 after this.

=======
Subject: fix split_huge_page error like mapcount 3 page_mapcount 2

From: Andrea Arcangeli <aarcange@redhat.com>

split_huge_page is very strict in checking the rmap walk was perfect. I'd
rather keep it strict because if page_mapcount isn't stable and just right, the
__split_huge_page_refcount that follows the rmap walk could lead to erratic
page_count()s for the subpages.

The below bug in fork lead to the rmap walk finding the parent huge-pmd twice
instead of just one, because the anon_vma_chain objects of the child vma still
point to the vma->vm_mm of the parent. That trips on the split_huge_page
mapcount vs page_mapcount check leading to a BUG_ON. The below patch fixes it
by making the rmap walk accurate during fork.

Signed-off-by: Andrea Arcangeli <aarcange@redhat.com>